requirements.txt
# PHANTOM Netrunner v3.0 - Dependencies
# ======================================

# Core Dependencies
requests>=2.28.0
pyyaml>=6.0

# Enhanced Terminal UI
rich>=13.0.0

# Web Framework (for social engineering module)
flask>=2.3.0

# Network Analysis
scapy>=2.5.0

# QR Code Generation
qrcode[pil]>=7.4

# Optional: Wireless Tools (system-level)
# Install via apt: aircrack-ng, reaver, wash

# Optional: Legacy Support
lolcat
boxes

# Development Dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
